requirements.txt

fastapi
uvicorn
paho-mqtt
